fix linters errors
